requirements.txt.txt
fastapi
uvicorn[standard]
python-dotenv
requests
faiss-cpu
numpy
sentence-transformers